задаёт сброс главной метки и немного рефакторит map.js и data.js
