+++ js/data.js
@@ -1,22 +1,6 @@
 const GET_DATA_URL = 'https://22.javascript.pages.academy/keksobooking/data';
 const POST_DATA_URL = 'https://22.javascript.pages.academy/keksobooking';
 
-const TARGET_AREA = {
-  startPoint: {
-    latitude: 35.65,
-    longitude: 139.7,
-  },
-  endPoint: {
-    latitude: 35.7,
-    longitude: 139.8,
-  },
-};
-
-let mainPoint = {
-  latitude: null,
-  longitude: null,
-};
-
 const HOUSING_TYPES = {
   'palace': {
     'caption': 'Дворец',
@@ -36,24 +20,7 @@ const HOUSING_TYPES = {
   },
 };
 
-const getAreaCenter = () => {
-  return {
-    latitude: (TARGET_AREA.startPoint.latitude + TARGET_AREA.endPoint.latitude) / 2,
-    longitude: (TARGET_AREA.startPoint.longitude + TARGET_AREA.endPoint.longitude) / 2,
-  };
-};
-
-const getMainPoint = () => {
-  if (!mainPoint.latitude) {
-    mainPoint = getAreaCenter();
-  }
-  return mainPoint;
-};
-
-const setMainPoint = ({ latitude, longitude }) => {
-  mainPoint.latitude = latitude;
-  mainPoint.longitude = longitude;
-};
+let similarNearAds = null;
 
 const getHousingCaption = (value) => {
   return HOUSING_TYPES[value]['caption'];
@@ -62,8 +29,6 @@ const getHousingCaption = (value) => {
 const getHousingMinPrice = (value) => {
   return HOUSING_TYPES[value]['minPrice'];
 };
-
-let similarNearAds = null;
 
 const fetchData = (succesHandler, errorHandler) => {
   fetch(GET_DATA_URL)
@@ -107,4 +72,4 @@ const getData = () => {
   return similarNearAds;
 };
 
-export { getHousingCaption, getHousingMinPrice, getMainPoint, setMainPoint, fetchData, postData, getData };
+export { getHousingCaption, getHousingMinPrice, fetchData, postData, getData };
